added support to pytorch dataloaders
