+++ apps/black_box_model_optimization/__init__.py
@@ -1,0 +1,3 @@
+from .black_box_model_optimization import (  # noqa F401
+    BlackBoxModelOptimization,
+)
