--- nebullvm/operations/root/black_box/utils.py
@@ -1,128 +1,0 @@
-import logging
-from pathlib import Path
-from typing import (
-    Dict,
-    Callable,
-    Union,
-    Iterable,
-    Sequence,
-    Any,
-    Optional,
-    List,
-)
-
-from nebullvm.optional_modules.tensorflow import tensorflow as tf
-from nebullvm.optional_modules.torch import Module
-from nebullvm.tools.base import DeepLearningFramework, Device, ModelParams
-from nebullvm.tools.data import DataManager
-from nebullvm.tools.onnx import (
-    extract_info_from_np_data,
-    get_output_sizes_onnx,
-)
-from nebullvm.tools.pytorch import (
-    extract_info_from_torch_data,
-    get_outputs_sizes_torch,
-)
-from nebullvm.tools.tensorflow import (
-    extract_info_from_tf_data,
-    get_outputs_sizes_tf,
-)
-
-logger = logging.getLogger("nebullvm_logger")
-
-INFO_EXTRACTION_DICT: Dict[DeepLearningFramework, Callable] = {
-    DeepLearningFramework.PYTORCH: extract_info_from_torch_data,
-    DeepLearningFramework.TENSORFLOW: extract_info_from_tf_data,
-    DeepLearningFramework.NUMPY: extract_info_from_np_data,
-}
-
-OUTPUT_SIZE_COMPUTATION_DICT: Dict[DeepLearningFramework, Callable] = {
-    DeepLearningFramework.PYTORCH: get_outputs_sizes_torch,
-    DeepLearningFramework.TENSORFLOW: get_outputs_sizes_tf,
-    DeepLearningFramework.NUMPY: get_output_sizes_onnx,
-}
-
-
-def check_input_data(input_data: Union[Iterable, Sequence]):
-    try:
-        input_data[0]
-    except:  # noqa E722
-        return False
-    else:
-        return True
-
-
-def extract_info_from_data(
-    model: Any,
-    input_data: DataManager,
-    dl_framework: DeepLearningFramework,
-    dynamic_info: Optional[Dict],
-    device: Device,
-):
-    batch_size, input_sizes, input_types, dynamic_info = INFO_EXTRACTION_DICT[
-        dl_framework
-    ](
-        model,
-        input_data,
-        batch_size=None,
-        input_sizes=None,
-        input_types=None,
-        dynamic_axis=dynamic_info,
-        device=device,
-    )
-    model_params = ModelParams(
-        batch_size=batch_size,
-        input_infos=[
-            {"size": size, "dtype": dtype}
-            for size, dtype in zip(input_sizes, input_types)
-        ],
-        output_sizes=OUTPUT_SIZE_COMPUTATION_DICT[dl_framework](
-            model, input_data[0][0], device
-        ),
-        dynamic_info=dynamic_info,
-    )
-    return model_params
-
-
-def get_dl_framework(model: Any):
-    if isinstance(model, Module):
-        return DeepLearningFramework.PYTORCH
-    elif isinstance(model, tf.Module) and model is not None:
-        return DeepLearningFramework.TENSORFLOW
-    elif isinstance(model, str):
-        if Path(model).is_file():
-            return DeepLearningFramework.NUMPY
-        else:
-            raise FileNotFoundError(
-                f"No file '{model}' found, please provide a valid path to "
-                f"a model."
-            )
-    else:
-        raise TypeError(f"Model type {type(model)} not supported.")
-
-
-def map_compilers_and_compressors(ignore_list: List, enum_class: Callable):
-    if ignore_list is None:
-        ignore_list = []
-    else:
-        ignore_list = [enum_class(element) for element in ignore_list]
-    return ignore_list
-
-
-def is_huggingface_data(data_sample: Any) -> bool:
-    if is_dict_type(data_sample):
-        return True
-    elif isinstance(data_sample, str):
-        return True
-    elif isinstance(data_sample[0], str):
-        return True
-    return False
-
-
-def is_dict_type(data_sample: Any):
-    try:
-        data_sample.items()
-    except AttributeError:
-        return False
-    else:
-        return True
