+++ nebullvm/optional_modules/tensorflow.py
@@ -1,3 +1,7 @@
+import absl.logging
+
+absl.logging.set_verbosity(absl.logging.ERROR)
+
 NoneType = type(None)
 
 
@@ -17,11 +21,17 @@ class Tensorflow:
 
 try:
     import tensorflow  # noqa F401
+
+    tensorflow.get_logger().setLevel("ERROR")
+    tensorflow.autograph.set_verbosity(0)
 except ImportError:
     tensorflow = Tensorflow
 
 
 try:
     import tf2onnx  # noqa F401
+
+    tf2onnx.logging.set_level("ERROR")
+    tf2onnx.logging.set_tf_verbosity("ERROR")
 except ImportError:
     tf2onnx = object
